isolate BlankBlock related logic
